+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## UNRELEASED (0.14.2)
+
+👌 **Improvements**
+
+- `BigInteger::modInverse()` now accepts `BigNumber|int|float|string` instead of just `BigInteger`
+
 ## [0.14.1](https://github.com/brick/math/releases/tag/0.14.1) - 2025-11-24
 
 ✨ **New features**